refactor(frontend): Move services in a dedicated folder.
Update references
