+++ frontend/src/app/footer/footer.component.ts
@@ -1,7 +1,6 @@
 import { Component, signal } from '@angular/core';
 import { faGithub, faLinkedin } from '@fortawesome/free-brands-svg-icons';
 import { faGlobe, faEnvelope } from '@fortawesome/free-solid-svg-icons';
-import { on } from 'events';
 import { environment } from 'src/environments/environment';
 
 @Component({
