+++ frontend/src/app/services/GlobalConfiguration/global-configuration.service.spec.ts
