+++ frontend/src/app/services/GlobalConfiguration/global-configuration.service.ts
