+++ frontend/src/app/services/VaultService/vault.service.spec.ts
