+++ frontend/src/app/services/VaultService/vault.service.ts
