+++ frontend/src/app/services/qr-code-totp/qr-code-totp.service.ts
